requirements.txt
Django~=5.0.6
djangorestframework~=3.15.2
asgiref==3.8.1
django-allauth==0.63.3
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==70.2.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
drf-spectacular==0.27.1
django-rest-swagger==2.2.0
